Fix compiler warnings: cast loses precision.
Use explicit cast when converting large integer types to smaller integer types when it is known that the most significant bytes are not required.
For pcidevice, check the ROM file size before casting to int. We'll allow expansion ROM sizes up to 4MB but usually they are 64K, sometimes 128K, rarely 256K.
for machinefactory, change the type to size_t so that it can correctly get the size of files that are larger than 4GB; it already checks the file size is 4MB before we need to cast to uint32_t.
For floppyimg, check the image size before casting to int. For raw images, only allow files up to 2MB. For DiskCopy42 images, it already checks the file size, so do the cast after that.

+++ cpu/ppc/ppcopcodes.cpp
@@ -729,7 +729,7 @@ void dppc_interpreter::ppc_srawi() {
 /** mask generator for rotate and shift instructions (§ 4.2.1.4 PowerpC PEM) */
 static inline uint32_t rot_mask(unsigned rot_mb, unsigned rot_me) {
     uint32_t m1 = 0xFFFFFFFFUL >> rot_mb;
-    uint32_t m2 = 0xFFFFFFFFUL << (31 - rot_me);
+    uint32_t m2 = (uint32_t)(0xFFFFFFFFUL << (31 - rot_me));
     return ((rot_mb <= rot_me) ? m2 & m1 : m1 | m2);
 }
 
@@ -861,11 +861,11 @@ static inline void calc_rtcl_value()
     uint64_t new_ts = get_virt_time_ns();
     uint64_t rtc_l = new_ts - rtc_timestamp + rtc_lo;
     if (rtc_l >= ONE_BILLION_NS) { // check RTCL overflow
-        rtc_hi += rtc_l / ONE_BILLION_NS;
+        rtc_hi += (uint32_t)(rtc_l / ONE_BILLION_NS);
         rtc_lo  = rtc_l % ONE_BILLION_NS;
     }
     else {
-        rtc_lo = rtc_l;
+        rtc_lo = (uint32_t)rtc_l;
     }
     rtc_timestamp = new_ts;
 }
